[step3] refactor: 계산기 패키지명 변경
