+++ src/test/java/calculator/CalculatorTest.java
@@ -1,4 +1,4 @@
-package step1;
+package calculator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
