+++ src/test/java/calculator/enums/OperationSymbolTest.java
@@ -1,4 +1,4 @@
-package step1.enums;
+package calculator.enums;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
