+++ src/test/java/calculator/validator/ValidatorTest.java
@@ -1,4 +1,4 @@
-package step1.validator;
+package calculator.validator;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.Test;
 import java.util.*;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.assertj.core.api.Assertions.assertThatCode;
 
 class ValidatorTest {
